# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_server_endpoints.AudioStatusResponse:
    properties:
      book_id:
        type: string
      chapter_count:
        type: integer
      chapters:
        items:
          $ref: '#/definitions/internal_server_endpoints.ChapterAudioStatus'
        type: array
      error_message:
        type: string
      format:
        type: string
      provider:
        type: string
      segment_count:
        type: integer
      status:
        type: string
      total_cost_usd:
        type: number
      total_duration_ms:
        type: integer
      voice:
        type: string
    type: object
  internal_server_endpoints.Book:
    properties:
      author:
        type: string
      created_at:
        type: string
      id:
        type: string
      metadata_complete:
        type: boolean
      page_count:
        type: integer
      status:
        type: string
      status_reason:
        type: string
      structure_complete:
        type: boolean
      structure_failed:
        type: boolean
      title:
        type: string
    type: object
...
  internal_server_endpoints.ChapterAudioStatus:
    properties:
      audio_file:
        type: string
      chapter_idx:
        type: integer
      cost_usd:
        type: number
      download_url:
        type: string
      duration_ms:
        type: integer
      segment_count:
        type: integer
      title:
        type: string
    type: object
  internal_server_endpoints.ChapterPage:
    properties:
      ocr_markdown:
        type: string
      page_num:
        type: integer
    type: object
  internal_server_endpoints.ChapterParagraph:
    properties:
      id:
        type: string
      polished_text:
        type: string
      raw_text:
        type: string
      sort_order:
        type: integer
      start_page:
        type: integer
      word_count:
        type: integer
    type: object
  internal_server_endpoints.ChapterWithText:
    properties:
      audio_include:
        type: boolean
      audio_include_reasoning:
        type: string
      classification_reasoning:
        type: string
      content_type:
        type: string
      edits_applied_json:
        type: string
      end_page:
        type: integer
      entry_id:
        type: string
      entry_number:
        type: string
      id:
        type: string
      level:
        type: integer
      level_name:
        type: string
      matter_type:
        type: string
      page_count:
        type: integer
      pages:
        items:
          $ref: '#/definitions/internal_server_endpoints.ChapterPage'
        type: array
      paragraphs:
        items:
          $ref: '#/definitions/internal_server_endpoints.ChapterParagraph'
        type: array
      polish_complete:
        type: boolean
      polish_failed:
        type: boolean
      polished_text:
        type: string
      sort_order:
        type: integer
      start_page:
        type: integer
      title:
        type: string
      word_count:
        type: integer
    type: object
  internal_server_endpoints.ChaptersResponse:
    properties:
      book_id:
        type: string
      book_title:
        type: string
      chapters:
        items:
          $ref: '#/definitions/internal_server_endpoints.ChapterWithText'
        type: array
      has_chapters:
        type: boolean
      total_pages:
        type: integer
    type: object
...
  internal_server_endpoints.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_server_endpoints.ListBooksResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/internal_server_endpoints.Book'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/jackzampolin/shelf
  description: Book digitization pipeline API for managing books, jobs, and processing.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Shelf API
  version: "1.0"
paths:
...
  /api/books:
    get:
      description: List all books in the library
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_endpoints.ListBooksResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: List books
      tags:
      - books
...
  /api/books/{book_id}/audio:
    get:
      description: Get TTS audio generation status and available files
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_endpoints.AudioStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Get audio status
      tags:
      - books
      - audio
...
  /api/books/{id}/chapters:
    get:
      description: Get chapter structure with page content
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Include page text content
        in: query
        name: include_text
        type: boolean
      - description: Include paragraphs
        in: query
        name: include_paragraphs
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_endpoints.ChaptersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Get book chapters
      tags:
      - books
...
swagger: "2.0"