# internal/server/endpoints/books_audio_status.go
package endpoints

import (
	"context"
	"fmt"
	"net/http"
	"sort"
	"strings"

	"github.com/spf13/cobra"

	"github.com/jackzampolin/shelf/internal/api"
	"github.com/jackzampolin/shelf/internal/defra"
	"github.com/jackzampolin/shelf/internal/jobs/tts_generate"
	"github.com/jackzampolin/shelf/internal/metrics"
	"github.com/jackzampolin/shelf/internal/svcctx"
)

// AudioStatusResponse contains audio generation status and files.
type AudioStatusResponse struct {
	BookID          string               `json:"book_id"`
	Status          string               `json:"status"`
	ErrorMessage    string               `json:"error_message,omitempty"`
	Provider        string               `json:"provider,omitempty"`
	Voice           string               `json:"voice,omitempty"`
	Format          string               `json:"format,omitempty"`
	TotalDurationMS int                  `json:"total_duration_ms,omitempty"`
	TotalCostUSD    float64              `json:"total_cost_usd,omitempty"`
	ChapterCount    int                  `json:"chapter_count,omitempty"`
	SegmentCount    int                  `json:"segment_count,omitempty"`
	Chapters        []ChapterAudioStatus `json:"chapters,omitempty"`
}

// ChapterAudioStatus contains status for a single chapter's audio.
type ChapterAudioStatus struct {
	ChapterIdx   int     `json:"chapter_idx"`
	Title        string  `json:"title,omitempty"`
	DurationMS   int     `json:"duration_ms,omitempty"`
	SegmentCount int     `json:"segment_count,omitempty"`
	CostUSD      float64 `json:"cost_usd,omitempty"`
	AudioFile    string  `json:"audio_file,omitempty"`
	DownloadURL  string  `json:"download_url,omitempty"`
}

// GetAudioStatusEndpoint handles GET /api/books/{book_id}/audio.
type GetAudioStatusEndpoint struct{}

func (e *GetAudioStatusEndpoint) Route() (string, string, http.HandlerFunc) {
	return "GET", "/api/books/{book_id}/audio", e.handler
}

func (e *GetAudioStatusEndpoint) RequiresInit() bool { return true }

// handler godoc
//
//	@Summary		Get audio status
//	@Description	Get TTS audio generation status and available files
//	@Tags			books,audio
//	@Produce		json
//	@Param			book_id	path		string	true	"Book ID"
//	@Success		200		{object}	AudioStatusResponse
//	@Failure		400		{object}	ErrorResponse
//	@Failure		404		{object}	ErrorResponse
//	@Failure		500		{object}	ErrorResponse
//	@Router			/api/books/{book_id}/audio [get]
func (e *GetAudioStatusEndpoint) handler(w http.ResponseWriter, r *http.Request) {
	bookID := r.PathValue("book_id")
	if bookID == "" {
		writeError(w, http.StatusBadRequest, "book_id is required")
		return
	}

	if err := defra.ValidateID(bookID); err != nil {
		writeError(w, http.StatusBadRequest, fmt.Sprintf("invalid book_id: %v", err))
		return
	}

	ctx := r.Context()
	defraClient := svcctx.DefraClientFrom(ctx)
	if defraClient == nil {
		writeError(w, http.StatusServiceUnavailable, "defra client not initialized")
		return
	}

	// Query BookAudio record
	// Note: Use unique_key (which stores the book's DocID) for lookups,
	// not book_id (which is auto-generated by DefraDB for the relationship).
	bookAudioQuery := fmt.Sprintf(`{
		BookAudio(filter: {unique_key: {_eq: "%s"}}) {
			_docID
			status
			error_message
			provider
			voice
			format
			total_duration_ms
			total_cost_usd
			chapter_count
			segment_count
		}
	}`, bookID)

	bookAudioResp, err := defraClient.Execute(ctx, bookAudioQuery, nil)
	if err != nil {
		writeError(w, http.StatusInternalServerError, fmt.Sprintf("failed to query audio status: %v", err))
		return
	}

	records, ok := bookAudioResp.Data["BookAudio"].([]any)
	if !ok || len(records) == 0 {
		writeJSON(w, http.StatusOK, AudioStatusResponse{
			BookID: bookID,
			Status: "not_started",
		})
		return
	}

	data, ok := records[0].(map[string]any)
	if !ok {
		writeError(w, http.StatusInternalServerError, "invalid audio data")
		return
	}

	resp := AudioStatusResponse{
		BookID:          bookID,
		Status:          getString(data, "status"),
		ErrorMessage:    getString(data, "error_message"),
		Provider:        getString(data, "provider"),
		Voice:           getString(data, "voice"),
		Format:          getString(data, "format"),
		TotalDurationMS: getInt(data, "total_duration_ms"),
		TotalCostUSD:    getFloat(data, "total_cost_usd"),
		ChapterCount:    getInt(data, "chapter_count"),
		SegmentCount:    getInt(data, "segment_count"),
	}

	// Query ChapterAudio records (prefer relation field, fall back to unique_key prefix).
	chapterRecords, err := queryChapterAudioRecords(ctx, defraClient, bookID)
	if err != nil {
		if logger := svcctx.LoggerFrom(ctx); logger != nil {
			logger.Warn("failed to query chapter audio records", "book_id", bookID, "error", err)
		}
	} else {
		chapterDocIDByIdx := make(map[int]string, len(chapterRecords))
		for _, chData := range chapterRecords {
			chapterIdx := getInt(chData, "chapter_idx")
			status := ChapterAudioStatus{
				ChapterIdx:   chapterIdx,
				DurationMS:   getInt(chData, "duration_ms"),
				SegmentCount: getInt(chData, "segment_count"),
				CostUSD:      getFloat(chData, "total_cost_usd"),
				AudioFile:    getString(chData, "audio_file"),
			}
			if status.AudioFile != "" {
				status.DownloadURL = fmt.Sprintf("/api/books/%s/audio/%d/download", bookID, chapterIdx)
			}
			if chapterDocID := chapterDocIDFromChapterAudioUniqueKey(getString(chData, "unique_key"), bookID); chapterDocID != "" {
				chapterDocIDByIdx[chapterIdx] = chapterDocID
			}
			resp.Chapters = append(resp.Chapters, status)
		}

		// Prefer metrics as source-of-truth for audiobook cost aggregation.
		metricsQuery := svcctx.MetricsQueryFrom(ctx)
		if metricsQuery == nil {
			metricsQuery = metrics.NewQuery(defraClient)
		}
		metricsTotalCost, metricsByChapter, err := audiobookCostsFromMetrics(ctx, metricsQuery, bookID, chapterDocIDByIdx)
		if err != nil {
			if logger := svcctx.LoggerFrom(ctx); logger != nil {
				logger.Warn("failed to query audiobook cost from metrics", "book_id", bookID, "error", err)
			}
		} else {
			resp.TotalCostUSD = metricsTotalCost
			for i := range resp.Chapters {
				if cost, ok := metricsByChapter[resp.Chapters[i].ChapterIdx]; ok {
					resp.Chapters[i].CostUSD = cost
				}
			}
		}
	}

	// Sort chapters by index
	sortChapterStatuses(resp.Chapters)

	writeJSON(w, http.StatusOK, resp)
}

func (e *GetAudioStatusEndpoint) Command(getServerURL func() string) *cobra.Command {
	return &cobra.Command{
		Use:   "audio <book_id>",
		Short: "Get audio status and files",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) error {
			ctx := cmd.Context()
			bookID := args[0]

			client := api.NewClient(getServerURL())
			var resp AudioStatusResponse
			if err := client.Get(ctx, fmt.Sprintf("/api/books/%s/audio", bookID), &resp); err != nil {
				return err
			}

			return api.Output(resp)
		},
	}
}

func sortChapterStatuses(chapters []ChapterAudioStatus) {
	sort.Slice(chapters, func(i, j int) bool {
		return chapters[i].ChapterIdx < chapters[j].ChapterIdx
	})
}

func getFloat(m map[string]any, key string) float64 {
	if v, ok := m[key].(float64); ok {
		return v
	}
	return 0
}

func chapterDocIDFromChapterAudioUniqueKey(uniqueKey, bookID string) string {
	if uniqueKey == "" {
		return ""
	}
	prefix := bookID + ":"
	if !strings.HasPrefix(uniqueKey, prefix) {
		return ""
	}
	rest := strings.TrimPrefix(uniqueKey, prefix)
	if rest == "" {
		return ""
	}
	parts := strings.SplitN(rest, ":", 2)
	return parts[0]
}

func chapterDocIDFromMetricItemKey(itemKey string) string {
	const marker = "_para_"
	idx := strings.Index(itemKey, marker)
	if idx <= 0 {
		return ""
	}
	return itemKey[:idx]
}

func audiobookCostsFromMetrics(ctx context.Context, query *metrics.Query, bookID string, chapterDocIDByIdx map[int]string) (float64, map[int]float64, error) {
	if query == nil {
		return 0, nil, nil
	}

	chapterIdxByDocID := make(map[string]int, len(chapterDocIDByIdx))
	for chapterIdx, chapterDocID := range chapterDocIDByIdx {
		if chapterDocID == "" {
			continue
		}
		chapterIdxByDocID[chapterDocID] = chapterIdx
	}

	stageNames := []string{tts_generate.JobTypeElevenLabs, tts_generate.JobTypeOpenAI}
	costByChapter := make(map[int]float64, len(chapterDocIDByIdx))
	total := 0.0

	for _, stage := range stageNames {
		metricsList, err := query.List(ctx, metrics.Filter{
			BookID: bookID,
			Stage:  stage,
		}, 0)
		if err != nil {
			return 0, nil, err
		}
		for _, metric := range metricsList {
			total += metric.CostUSD
			chapterDocID := chapterDocIDFromMetricItemKey(metric.ItemKey)
			chapterIdx, ok := chapterIdxByDocID[chapterDocID]
			if !ok {
				continue
			}
			costByChapter[chapterIdx] += metric.CostUSD
		}
	}

	return total, costByChapter, nil
}